Mobile login: Fixed forward to response page.
* Show user name before linking accounts.
