+++ phoneblock/src/main/java/de/haumacher/phoneblock/app/render/TemplateRenderer.java
@@ -10,6 +10,7 @@ import org.thymeleaf.web.servlet.JakartaServletWebApplication;
 
 import de.haumacher.phoneblock.app.SettingsServlet;
 import de.haumacher.phoneblock.app.render.controller.LoginController;
+import de.haumacher.phoneblock.app.render.controller.MobileLoginController;
 import de.haumacher.phoneblock.app.render.controller.PhoneInfoController;
 import de.haumacher.phoneblock.app.render.controller.RequireLoginController;
 import de.haumacher.phoneblock.app.render.controller.SettingsController;
@@ -27,7 +28,7 @@ public class TemplateRenderer {
     static {
         controllersByURL = new HashMap<String, WebController>();
         controllersByURL.put("/login", new LoginController());
-        controllersByURL.put("/mobile/login", new LoginController());
+        controllersByURL.put("/mobile/login", new MobileLoginController());
         controllersByURL.put("/status", new StatusController());
         controllersByURL.put(SettingsServlet.PATH, new SettingsController());
         controllersByURL.put("/phone-info", new PhoneInfoController());
