+++ phoneblock/src/main/java/de/haumacher/phoneblock/app/render/controller/LoginController.java
@@ -16,7 +16,7 @@ import de.haumacher.phoneblock.random.SecureRandomService;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 
-public class LoginController extends DefaultController {
+public class LoginController extends MobileLoginController {
 	
 	@Override
 	public boolean process(TemplateRenderer renderer, HttpServletRequest request, HttpServletResponse response) throws IOException {
@@ -30,26 +30,4 @@ public class LoginController extends DefaultController {
 		return super.process(renderer, request, response);
 	}
 	
-	@Override
-	protected void fillContext(WebContext ctx, HttpServletRequest request) {
-		super.fillContext(ctx, request);
-		
-		ctx.setVariable("googleClient", PhoneBlockConfigFactory.GOOGLE_CLIENT);
-		ctx.setVariable("loginWeb", EMailVerificationServlet.LOGIN_WEB);
-		ctx.setVariable("loginMobile", EMailVerificationServlet.LOGIN_MOBILE);
-		ctx.setVariable("loginAction", LoginServlet.PATH);
-		ctx.setVariable("rememberParam", LoginServlet.REMEMBER_ME_PARAM);
-		ctx.setVariable("userNameParam", LoginServlet.USER_NAME_PARAM);
-		ctx.setVariable("passwordParam", LoginServlet.PASSWORD_PARAM);
-		
-		addCaptcha(ctx, request);
-	}
-
-	public static void addCaptcha(WebContext ctx, HttpServletRequest request) {
-		Captcha captcha = new Captcha(SecureRandomService.getInstance().getRnd());
-		request.getSession().setAttribute("captcha", captcha.getText());
-		
-		ctx.setVariable("captchaSrc", "data:image/png;base64, " + Base64.getEncoder().encodeToString(captcha.getPng()));
-	}
-
 }
