+++ phoneblock/src/main/webapp/WEB-INF/properties/ar/mobile/login.properties
@@ -2,4 +2,5 @@ t0001=\u0631\u0628\u0637 \u062d\u0633\u0627\u0628 PhoneBlock
 t0002=\u0647\u0646\u0627 \u062a\u0642\u0648\u0645 \u0628\u0631\u0628\u0637 \u062d\u0633\u0627\u0628 PhoneBlock \u0627\u0644\u062e\u0627\u0635 \u0628\u0643 \u0645\u0639 \u062c\u0647\u0627\u0632\u0643 \u0627\u0644\u0645\u062d\u0645\u0648\u0644.
 t0003=\u064a\u062c\u0628 \u0639\u0644\u064a\u0643 \u0623\u0648\u0644\u0627\u064b \u0627\u0644\u062a\u0633\u062c\u064a\u0644 \u0641\u064a PhoneBlock \u0623\u0648 \u0625\u0646\u0634\u0627\u0621 \u062d\u0633\u0627\u0628 PhoneBlock. \u062a\u062d\u062a\u0627\u062c \u0625\u0644\u0649 \u062d\u0633\u0627\u0628 PhoneBlock \u062d\u062a\u0649 \u064a\u062a\u0645\u0643\u0646 \u0647\u0627\u062a\u0641\u0643 \u0627\u0644\u0645\u062d\u0645\u0648\u0644 \u0645\u0646 \u0627\u0633\u062a\u062e\u062f\u0627\u0645 \u0642\u0627\u0639\u062f\u0629 \u0628\u064a\u0627\u0646\u0627\u062a PhoneBlock \u0644\u062a\u0635\u0641\u064a\u0629 \u0627\u0644\u0645\u0643\u0627\u0644\u0645\u0627\u062a.
 t0004=\u0631\u0628\u0637 \u062d\u0633\u0627\u0628 PhoneBlock
+t0005=\u0644\u0642\u062f \u0642\u0645\u062a \u0628\u062a\u0633\u062c\u064a\u0644 \u0627\u0644\u062f\u062e\u0648\u0644 \u0625\u0644\u0649 PhoneBlock \u0628\u0627\u0633\u0645 <x1></x1>.
