+++ phoneblock/src/main/webapp/WEB-INF/properties/ar/mobile/response.properties
@@ -1,4 +1,3 @@
 t0001=\u062a\u0645 \u062a\u0633\u062c\u064a\u0644 \u062f\u062e\u0648\u0644\u0643 \u0628\u0646\u062c\u0627\u062d
-t0002=\u0631\u0645\u0632 \u0627\u0644\u062a\u0641\u0648\u064a\u0636 \u0627\u0644\u062e\u0627\u0635 \u0628\u0643: <x1></x1>
 t0003=\u062a\u0645 \u062a\u0633\u062c\u064a\u0644 \u062f\u062e\u0648\u0644\u0643 \u0625\u0644\u0649 PhoneBlock \u0628\u0646\u062c\u0627\u062d. \u0625\u0630\u0627 \u0644\u0645 \u064a\u062a\u0645 \u0625\u063a\u0644\u0627\u0642 \u0647\u0630\u0647 \u0627\u0644\u0635\u0641\u062d\u0629 \u062a\u0644\u0642\u0627\u0626\u064a\u064b\u0627\u060c \u0641\u0642\u062f \u064a\u0643\u0648\u0646 \u0647\u0646\u0627\u0643 \u062e\u0637\u0623 \u0645\u0627 \u0641\u064a \u062a\u0637\u0628\u064a\u0642 \u0647\u0627\u062a\u0641\u0643 \u0627\u0644\u0645\u062d\u0645\u0648\u0644.
