+++ phoneblock/src/main/webapp/WEB-INF/properties/da/mobile/login.properties
@@ -2,4 +2,5 @@ t0001=Link til PhoneBlock-konto
 t0002=Her forbinder du din PhoneBlock-konto med din mobile enhed.
 t0003=Du skal først registrere dig hos PhoneBlock eller oprette en PhoneBlock-konto. Du skal have en PhoneBlock-konto, så din mobiltelefon kan bruge PhoneBlock-databasen til at filtrere opkald.
 t0004=Link til PhoneBlock-konto
+t0005=Du er logget ind på PhoneBlock som <x1></x1>.
