+++ phoneblock/src/main/webapp/WEB-INF/properties/da/mobile/response.properties
@@ -1,4 +1,3 @@
 t0001=Du er logget ind med succes
-t0002=Dit autorisationstoken: <x1></x1>.
 t0003=Dit login til PhoneBlock var vellykket. Hvis denne side ikke lukker automatisk, kan der være noget galt med din mobilapplikation.
