+++ phoneblock/src/main/webapp/WEB-INF/properties/de/mobile/login.properties
@@ -2,4 +2,5 @@ t0001=PhoneBlock-Account verknüpfen
 t0002=Hier verbindest du Deinen PhoneBlock-Account mit Deinem Mobilgerät.
 t0003=Du musst Dich zuerst bei PhoneBlock anmelden bzw. einen PhoneBlock-Account erstellen. Du benötigst einen PhoneBlock-Account, damit Dein Mobiltelefon die PhoneBlock-Datenbank für die Filterung von Anrufen nutzen kann.
 t0004=PhoneBlock-Account verknüpfen
+t0005=Du bist als <x1></x1> bei PhoneBlock angemeldet.
