+++ phoneblock/src/main/webapp/WEB-INF/properties/de/mobile/response.properties
@@ -1,4 +1,3 @@
 t0001=Du bist erfolgreich angemeldet
-t0002=Dein Authorisierungstoken: <x1></x1>
 t0003=Deine Anmeldung an PhoneBlock war erfolgreich. Wenn sich diese Seite nicht automatisch schließt ist möglicherweise etwas mit Deiner mobilen Anwendung nicht in Ordnung.
