+++ phoneblock/src/main/webapp/WEB-INF/properties/en-US/mobile/login.properties
@@ -2,4 +2,5 @@ t0001=Link PhoneBlock account
 t0002=Here you connect your PhoneBlock account with your mobile device.
 t0003=You must first register with PhoneBlock or create a PhoneBlock account. You need a PhoneBlock account so that your cell phone can use the PhoneBlock database to filter calls.
 t0004=Link PhoneBlock account
+t0005=You are logged in to PhoneBlock as <x1></x1>.
