+++ phoneblock/src/main/webapp/WEB-INF/properties/en-US/mobile/response.properties
@@ -1,4 +1,3 @@
 t0001=You are successfully logged in
-t0002=Your authorization token: <x1></x1>
 t0003=Your login to PhoneBlock was successful. If this page does not close automatically, there may be something wrong with your mobile application.
