+++ phoneblock/src/main/webapp/WEB-INF/properties/es/mobile/login.properties
@@ -2,4 +2,5 @@ t0001=Vincular cuenta PhoneBlock
 t0002=Aquí conectas tu cuenta PhoneBlock con tu dispositivo móvil.
 t0003=Primero debe registrarse en PhoneBlock o crear una cuenta PhoneBlock. Necesita una cuenta PhoneBlock para que su teléfono móvil pueda utilizar la base de datos PhoneBlock para filtrar las llamadas.
 t0004=Vincular cuenta PhoneBlock
+t0005=Has iniciado sesión en PhoneBlock como <x1></x1>.
