+++ phoneblock/src/main/webapp/WEB-INF/properties/es/mobile/response.properties
@@ -1,4 +1,3 @@
 t0001=Ha iniciado sesión correctamente
-t0002=Su token de autorización: <x1></x1>
 t0003=Su acceso a PhoneBlock se ha realizado correctamente. Si esta página no se cierra automáticamente, es posible que haya algún problema con su aplicación móvil.
