+++ phoneblock/src/main/webapp/WEB-INF/properties/fr/mobile/login.properties
@@ -2,4 +2,5 @@ t0001=Associer un compte PhoneBlock
 t0002=Ici, tu connectes ton compte PhoneBlock à ton appareil mobile.
 t0003=Tu dois d'abord t'inscrire sur PhoneBlock ou créer un compte PhoneBlock. Tu as besoin d'un compte PhoneBlock pour que ton téléphone portable puisse utiliser la base de données PhoneBlock pour filtrer les appels.
 t0004=Associer un compte PhoneBlock
+t0005=Tu es connecté à PhoneBlock en tant que <x1></x1>.
