+++ phoneblock/src/main/webapp/WEB-INF/properties/fr/mobile/response.properties
@@ -1,4 +1,3 @@
 t0001=Tu es inscrit(e) avec succès
-t0002=Ton jeton d'autorisation : <x1></x1>
 t0003=Ton inscription à PhoneBlock a réussi. Si cette page ne se ferme pas automatiquement, il se peut que quelque chose ne fonctionne pas avec ton application mobile.
