+++ phoneblock/src/main/webapp/WEB-INF/properties/it/mobile/login.properties
@@ -2,4 +2,5 @@ t0001=Collegare l'account PhoneBlock
 t0002=Qui si collega l'account PhoneBlock al dispositivo mobile.
 t0003=È necessario prima registrarsi a PhoneBlock o creare un account PhoneBlock. È necessario un account PhoneBlock affinché il telefono cellulare possa utilizzare il database PhoneBlock per filtrare le chiamate.
 t0004=Collegare l'account PhoneBlock
+t0005=Siete collegati a PhoneBlock come <x1></x1>.
