+++ phoneblock/src/main/webapp/WEB-INF/properties/it/mobile/response.properties
@@ -1,4 +1,3 @@
 t0001=L'accesso è stato effettuato con successo
-t0002=Il suo token di autorizzazione: <x1></x1>
 t0003=Il vostro accesso a PhoneBlock è avvenuto con successo. Se questa pagina non si chiude automaticamente, potrebbe esserci qualcosa di sbagliato nell'applicazione mobile.
