+++ phoneblock/src/main/webapp/WEB-INF/properties/nb/mobile/login.properties
@@ -2,4 +2,5 @@ t0001=Koble til PhoneBlock-konto
 t0002=Her kobler du PhoneBlock-kontoen din til mobilenheten din.
 t0003=Du må først registrere deg hos PhoneBlock eller opprette en PhoneBlock-konto. Du trenger en PhoneBlock-konto for at mobiltelefonen din skal kunne bruke PhoneBlock-databasen til å filtrere anrop.
 t0004=Koble til PhoneBlock-konto
+t0005=Du er logget inn på PhoneBlock som <x1></x1>.
