+++ phoneblock/src/main/webapp/WEB-INF/properties/nb/mobile/response.properties
@@ -1,4 +1,3 @@
 t0001=Du er vellykket logget inn
-t0002=Ditt autorisasjonstoken: <x1></x1>.
 t0003=Innloggingen din til PhoneBlock var vellykket. Hvis denne siden ikke lukkes automatisk, kan det være noe galt med mobilapplikasjonen din.
