+++ phoneblock/src/main/webapp/WEB-INF/properties/nl/mobile/login.properties
@@ -2,4 +2,5 @@ t0001=PhoneBlock-account koppelen
 t0002=Hier verbindt u uw PhoneBlock-account met uw mobiele apparaat.
 t0003=U moet zich eerst registreren bij PhoneBlock of een PhoneBlock-account aanmaken. U hebt een PhoneBlock account nodig zodat uw mobiele telefoon de PhoneBlock database kan gebruiken om oproepen te filteren.
 t0004=PhoneBlock-account koppelen
+t0005=U bent aangemeld bij PhoneBlock als <x1></x1>.
