+++ phoneblock/src/main/webapp/WEB-INF/properties/nl/mobile/response.properties
@@ -1,4 +1,3 @@
 t0001=U bent succesvol ingelogd
-t0002=Uw autorisatietoken: <x1></x1>
 t0003=Uw aanmelding bij PhoneBlock is gelukt. Als deze pagina niet automatisch sluit, is er mogelijk iets mis met uw mobiele applicatie.
