+++ phoneblock/src/main/webapp/WEB-INF/properties/pl/mobile/login.properties
@@ -2,4 +2,5 @@ t0001=Po\u0142\u0105cz konto PhoneBlock
 t0002=Tutaj \u0142\u0105czysz swoje konto PhoneBlock z urz\u0105dzeniem mobilnym.
 t0003=Najpierw nale\u017cy zarejestrowa\u0107 si\u0119 w us\u0142udze PhoneBlock lub utworzy\u0107 konto PhoneBlock. Konto PhoneBlock jest potrzebne, aby telefon komórkowy móg\u0142 korzysta\u0107 z bazy danych PhoneBlock do filtrowania po\u0142\u0105cze\u0144.
 t0004=Po\u0142\u0105cz konto PhoneBlock
+t0005=Jeste\u015b zalogowany do PhoneBlock jako <x1></x1>.
