+++ phoneblock/src/main/webapp/WEB-INF/properties/pl/mobile/response.properties
@@ -1,4 +1,3 @@
 t0001=Zosta\u0142e\u015b pomy\u015blnie zalogowany
-t0002=Twój token autoryzacji: <x1></x1>
 t0003=Logowanie do PhoneBlock powiod\u0142o si\u0119. Je\u015bli ta strona nie zamknie si\u0119 automatycznie, by\u0107 mo\u017ce co\u015b jest nie tak z aplikacj\u0105 mobiln\u0105.
