+++ phoneblock/src/main/webapp/WEB-INF/properties/sv/mobile/login.properties
@@ -2,4 +2,5 @@ t0001=Länka PhoneBlock-konto
 t0002=Här ansluter du ditt PhoneBlock-konto till din mobila enhet.
 t0003=Du måste först registrera dig hos PhoneBlock eller skapa ett PhoneBlock-konto. Du behöver ett PhoneBlock-konto för att din mobiltelefon ska kunna använda PhoneBlock-databasen för att filtrera samtal.
 t0004=Länka PhoneBlock-konto
+t0005=Du är inloggad på PhoneBlock som <x1></x1>.
