+++ phoneblock/src/main/webapp/WEB-INF/properties/sv/mobile/response.properties
@@ -1,4 +1,3 @@
 t0001=Du är framgångsrikt inloggad
-t0002=Din behörighetstoken: <x1></x1>.
 t0003=Din inloggning till PhoneBlock var framgångsrik. Om den här sidan inte stängs automatiskt kan det vara något fel på din mobilapplikation.
