+++ phoneblock/src/main/webapp/WEB-INF/properties/uk/mobile/login.properties
@@ -2,4 +2,5 @@ t0001=\u041f\u0440\u0438\u0432'\u044f\u0437\u0430\u0442\u0438 \u043e\u0431\u043b
 t0002=\u0422\u0443\u0442 \u0432\u0438 \u0437\u0432'\u044f\u0437\u0443\u0454\u0442\u0435 \u0441\u0432\u0456\u0439 \u043e\u0431\u043b\u0456\u043a\u043e\u0432\u0438\u0439 \u0437\u0430\u043f\u0438\u0441 PhoneBlock \u0437 \u0432\u0430\u0448\u0438\u043c \u043c\u043e\u0431\u0456\u043b\u044c\u043d\u0438\u043c \u043f\u0440\u0438\u0441\u0442\u0440\u043e\u0454\u043c.
 t0003=\u0421\u043f\u043e\u0447\u0430\u0442\u043a\u0443 \u0432\u0438 \u043f\u043e\u0432\u0438\u043d\u043d\u0456 \u0437\u0430\u0440\u0435\u0454\u0441\u0442\u0440\u0443\u0432\u0430\u0442\u0438\u0441\u044f \u0432 PhoneBlock \u0430\u0431\u043e \u0441\u0442\u0432\u043e\u0440\u0438\u0442\u0438 \u043e\u0431\u043b\u0456\u043a\u043e\u0432\u0438\u0439 \u0437\u0430\u043f\u0438\u0441 PhoneBlock. \u041e\u0431\u043b\u0456\u043a\u043e\u0432\u0438\u0439 \u0437\u0430\u043f\u0438\u0441 PhoneBlock \u043f\u043e\u0442\u0440\u0456\u0431\u0435\u043d \u0434\u043b\u044f \u0442\u043e\u0433\u043e, \u0449\u043e\u0431 \u0432\u0430\u0448 \u043c\u043e\u0431\u0456\u043b\u044c\u043d\u0438\u0439 \u0442\u0435\u043b\u0435\u0444\u043e\u043d \u043c\u0456\u0433 \u0432\u0438\u043a\u043e\u0440\u0438\u0441\u0442\u043e\u0432\u0443\u0432\u0430\u0442\u0438 \u0431\u0430\u0437\u0443 \u0434\u0430\u043d\u0438\u0445 PhoneBlock \u0434\u043b\u044f \u0444\u0456\u043b\u044c\u0442\u0440\u0430\u0446\u0456\u0457 \u0434\u0437\u0432\u0456\u043d\u043a\u0456\u0432.
 t0004=\u041f\u0440\u0438\u0432'\u044f\u0437\u0430\u0442\u0438 \u043e\u0431\u043b\u0456\u043a\u043e\u0432\u0438\u0439 \u0437\u0430\u043f\u0438\u0441 PhoneBlock
+t0005=\u0412\u0438 \u0443\u0432\u0456\u0439\u0448\u043b\u0438 \u0434\u043e PhoneBlock \u044f\u043a <x1></x1>.
