+++ phoneblock/src/main/webapp/WEB-INF/properties/uk/mobile/response.properties
@@ -1,4 +1,3 @@
 t0001=\u0412\u0438 \u0443\u0441\u043f\u0456\u0448\u043d\u043e \u0443\u0432\u0456\u0439\u0448\u043b\u0438 \u0434\u043e \u0441\u0438\u0441\u0442\u0435\u043c\u0438
-t0002=\u0412\u0430\u0448 \u0442\u043e\u043a\u0435\u043d \u0430\u0432\u0442\u043e\u0440\u0438\u0437\u0430\u0446\u0456\u0457: <x1></x1>
 t0003=\u0412\u0430\u0448 \u0432\u0445\u0456\u0434 \u0434\u043e PhoneBlock \u043f\u0440\u043e\u0439\u0448\u043e\u0432 \u0443\u0441\u043f\u0456\u0448\u043d\u043e. \u042f\u043a\u0449\u043e \u0446\u044f \u0441\u0442\u043e\u0440\u0456\u043d\u043a\u0430 \u043d\u0435 \u0437\u0430\u043a\u0440\u0438\u0432\u0430\u0454\u0442\u044c\u0441\u044f \u0430\u0432\u0442\u043e\u043c\u0430\u0442\u0438\u0447\u043d\u043e, \u043c\u043e\u0436\u043b\u0438\u0432\u043e, \u0437 \u0432\u0430\u0448\u0438\u043c \u043c\u043e\u0431\u0456\u043b\u044c\u043d\u0438\u043c \u0434\u043e\u0434\u0430\u0442\u043a\u043e\u043c \u0449\u043e\u0441\u044c \u043d\u0435 \u0442\u0430\u043a.
