+++ phoneblock/src/main/webapp/WEB-INF/properties/zh-Hans/mobile/login.properties
@@ -2,4 +2,5 @@ t0001=\u94fe\u63a5 PhoneBlock \u8d26\u6237
 t0002=\u5728\u8fd9\u91cc\uff0c\u60a8\u53ef\u4ee5\u5c06 PhoneBlock \u8d26\u6237\u4e0e\u79fb\u52a8\u8bbe\u5907\u8fde\u63a5\u8d77\u6765\u3002
 t0003=\u60a8\u5fc5\u987b\u9996\u5148\u5728 PhoneBlock \u6ce8\u518c\u6216\u521b\u5efa\u4e00\u4e2a PhoneBlock \u8d26\u6237\u3002\u60a8\u9700\u8981\u4e00\u4e2a PhoneBlock \u5e10\u6237\uff0c\u8fd9\u6837\u60a8\u7684\u624b\u673a\u624d\u80fd\u4f7f\u7528 PhoneBlock \u6570\u636e\u5e93\u8fc7\u6ee4\u6765\u7535\u3002
 t0004=\u94fe\u63a5 PhoneBlock \u8d26\u6237
+t0005=\u60a8\u6b63\u4ee5 <x1></x1> \u7684\u8eab\u4efd\u767b\u5f55 PhoneBlock\u3002
