+++ phoneblock/src/main/webapp/WEB-INF/properties/zh-Hans/mobile/response.properties
@@ -1,4 +1,3 @@
 t0001=\u60a8\u5df2\u6210\u529f\u767b\u5f55
-t0002=\u60a8\u7684\u6388\u6743\u4ee4\u724c\uff1a <x1></x1>
 t0003=\u60a8\u5df2\u6210\u529f\u767b\u5f55 PhoneBlock\u3002\u5982\u679c\u6b64\u9875\u9762\u6ca1\u6709\u81ea\u52a8\u5173\u95ed\uff0c\u5219\u53ef\u80fd\u662f\u60a8\u7684\u624b\u673a\u5e94\u7528\u7a0b\u5e8f\u51fa\u4e86\u95ee\u9898\u3002
