+++ phoneblock/src/main/webapp/WEB-INF/templates/ar/mobile/login.html
@@ -9,15 +9,19 @@
 	<div class="content">
 	<h1 data-tx="t0001">ربط حساب PhoneBlock</h1>
 	
+	<p data-tx="t0005" th:unless="${#strings.isEmpty(userName)}">لقد قمت بتسجيل الدخول إلى PhoneBlock باسم <code th:text="${userName}"></code>.</p>
+	
 	<p data-tx="t0002">هنا تقوم بربط حساب PhoneBlock الخاص بك مع جهازك المحمول.</p>
 
-	<p data-tx="t0003" th:if="${#strings.isEmpty(login)}">يجب عليك أولاً التسجيل في PhoneBlock أو إنشاء حساب PhoneBlock. تحتاج إلى حساب PhoneBlock حتى يتمكن هاتفك المحمول من استخدام قاعدة بيانات PhoneBlock لتصفية المكالمات.</p>
+	<p data-tx="t0003" th:if="${#strings.isEmpty(userName)}">يجب عليك أولاً التسجيل في PhoneBlock أو إنشاء حساب PhoneBlock. تحتاج إلى حساب PhoneBlock حتى يتمكن هاتفك المحمول من استخدام قاعدة بيانات PhoneBlock لتصفية المكالمات.</p>
 	</div>
 
-	<div th:if="${#strings.isEmpty(login)}" th:replace="~{../fragments/login-forms :: forms(web=false, location='/mobile/login')}">
+	<th:block th:if="${#strings.isEmpty(userName)}">
+	<div th:replace="~{../fragments/login-forms :: forms(web=false, location='/mobile/login')}">
 	</div>
+	</th:block>
 
-	<th:block th:unless="${#strings.isEmpty(login)}">
+	<th:block th:unless="${#strings.isEmpty(userName)}">
 	<form method="post" th:action="@{/create-token}">
 		<div class="field is-grouped">
 		  <p class="control">
