+++ phoneblock/src/main/webapp/WEB-INF/templates/ar/mobile/response.html
@@ -10,8 +10,6 @@
 
 <h1 data-tx="t0001">تم تسجيل دخولك بنجاح</h1>
 
-<p data-tx="t0002">رمز التفويض الخاص بك: <code id="token" th:text="${token}"></code></p>
-
 <p data-tx="t0003">تم تسجيل دخولك إلى PhoneBlock بنجاح. إذا لم يتم إغلاق هذه الصفحة تلقائيًا، فقد يكون هناك خطأ ما في تطبيق هاتفك المحمول.</p>
 
 </div>
