+++ phoneblock/src/main/webapp/WEB-INF/templates/da/mobile/login.html
@@ -9,15 +9,19 @@
 	<div class="content">
 	<h1 data-tx="t0001">Link til PhoneBlock-konto</h1>
 	
+	<p data-tx="t0005" th:unless="${#strings.isEmpty(userName)}">Du er logget ind på PhoneBlock som <code th:text="${userName}"></code>.</p>
+	
 	<p data-tx="t0002">Her forbinder du din PhoneBlock-konto med din mobile enhed.</p>
 
-	<p data-tx="t0003" th:if="${#strings.isEmpty(login)}">Du skal først registrere dig hos PhoneBlock eller oprette en PhoneBlock-konto. Du skal have en PhoneBlock-konto, så din mobiltelefon kan bruge PhoneBlock-databasen til at filtrere opkald.</p>
+	<p data-tx="t0003" th:if="${#strings.isEmpty(userName)}">Du skal først registrere dig hos PhoneBlock eller oprette en PhoneBlock-konto. Du skal have en PhoneBlock-konto, så din mobiltelefon kan bruge PhoneBlock-databasen til at filtrere opkald.</p>
 	</div>
 
-	<div th:if="${#strings.isEmpty(login)}" th:replace="~{../fragments/login-forms :: forms(web=false, location='/mobile/login')}">
+	<th:block th:if="${#strings.isEmpty(userName)}">
+	<div th:replace="~{../fragments/login-forms :: forms(web=false, location='/mobile/login')}">
 	</div>
+	</th:block>
 
-	<th:block th:unless="${#strings.isEmpty(login)}">
+	<th:block th:unless="${#strings.isEmpty(userName)}">
 	<form method="post" th:action="@{/create-token}">
 		<div class="field is-grouped">
 		  <p class="control">
