+++ phoneblock/src/main/webapp/WEB-INF/templates/da/mobile/response.html
@@ -10,8 +10,6 @@
 
 <h1 data-tx="t0001">Du er logget ind med succes</h1>
 
-<p data-tx="t0002">Dit autorisationstoken: <code id="token" th:text="${token}"></code>.</p>
-
 <p data-tx="t0003">Dit login til PhoneBlock var vellykket. Hvis denne side ikke lukker automatisk, kan der være noget galt med din mobilapplikation.</p>
 
 </div>
