+++ phoneblock/src/main/webapp/WEB-INF/templates/de/mobile/response.html
@@ -10,10 +10,6 @@
 
 <h1 data-tx="t0001">Du bist erfolgreich angemeldet</h1>
 
-<p data-tx="t0002">
-	Dein Authorisierungstoken: <code id="token" th:text="${token}"></code>
-</p>
-
 <p data-tx="t0003">
 	Deine Anmeldung an PhoneBlock war erfolgreich. Wenn sich diese Seite nicht automatisch schließt ist 
 	möglicherweise etwas mit Deiner mobilen Anwendung nicht in Ordnung.
