+++ phoneblock/src/main/webapp/WEB-INF/templates/el/mobile/login.html
@@ -9,15 +9,19 @@
 	<div class="content">
 	<h1 data-tx="t0001">Σύνδεση λογαριασμού PhoneBlock</h1>
 	
+	<p data-tx="t0005" th:unless="${#strings.isEmpty(userName)}">Έχετε συνδεθεί στο PhoneBlock ως <code th:text="${userName}"></code>.</p>
+	
 	<p data-tx="t0002">Εδώ συνδέετε το λογαριασμό σας PhoneBlock με την κινητή συσκευή σας.</p>
 
-	<p data-tx="t0003" th:if="${#strings.isEmpty(login)}">Πρέπει πρώτα να εγγραφείτε στο PhoneBlock ή να δημιουργήσετε έναν λογαριασμό PhoneBlock. Χρειάζεστε έναν λογαριασμό PhoneBlock για να μπορεί το κινητό σας τηλέφωνο να χρησιμοποιεί τη βάση δεδομένων PhoneBlock για να φιλτράρει τις κλήσεις.</p>
+	<p data-tx="t0003" th:if="${#strings.isEmpty(userName)}">Πρέπει πρώτα να εγγραφείτε στο PhoneBlock ή να δημιουργήσετε έναν λογαριασμό PhoneBlock. Χρειάζεστε έναν λογαριασμό PhoneBlock για να μπορεί το κινητό σας τηλέφωνο να χρησιμοποιεί τη βάση δεδομένων PhoneBlock για να φιλτράρει τις κλήσεις.</p>
 	</div>
 
-	<div th:if="${#strings.isEmpty(login)}" th:replace="~{../fragments/login-forms :: forms(web=false, location='/mobile/login')}">
+	<th:block th:if="${#strings.isEmpty(userName)}">
+	<div th:replace="~{../fragments/login-forms :: forms(web=false, location='/mobile/login')}">
 	</div>
+	</th:block>
 
-	<th:block th:unless="${#strings.isEmpty(login)}">
+	<th:block th:unless="${#strings.isEmpty(userName)}">
 	<form method="post" th:action="@{/create-token}">
 		<div class="field is-grouped">
 		  <p class="control">
