+++ phoneblock/src/main/webapp/WEB-INF/templates/el/mobile/response.html
@@ -10,8 +10,6 @@
 
 <h1 data-tx="t0001">Έχετε συνδεθεί επιτυχώς</h1>
 
-<p data-tx="t0002">Το διακριτικό εξουσιοδότησης: <code id="token" th:text="${token}"></code></p>
-
 <p data-tx="t0003">Η σύνδεσή σας στο PhoneBlock ήταν επιτυχής. Εάν αυτή η σελίδα δεν κλείσει αυτόματα, μπορεί να υπάρχει κάποιο πρόβλημα με την εφαρμογή κινητού σας.</p>
 
 </div>
