+++ phoneblock/src/main/webapp/WEB-INF/templates/en-US/mobile/login.html
@@ -9,15 +9,19 @@
 	<div class="content">
 	<h1 data-tx="t0001">Link PhoneBlock account</h1>
 	
+	<p data-tx="t0005" th:unless="${#strings.isEmpty(userName)}">You are logged in to PhoneBlock as <code th:text="${userName}"></code>.</p>
+	
 	<p data-tx="t0002">Here you connect your PhoneBlock account with your mobile device.</p>
 
-	<p data-tx="t0003" th:if="${#strings.isEmpty(login)}">You must first register with PhoneBlock or create a PhoneBlock account. You need a PhoneBlock account so that your cell phone can use the PhoneBlock database to filter calls.</p>
+	<p data-tx="t0003" th:if="${#strings.isEmpty(userName)}">You must first register with PhoneBlock or create a PhoneBlock account. You need a PhoneBlock account so that your cell phone can use the PhoneBlock database to filter calls.</p>
 	</div>
 
-	<div th:if="${#strings.isEmpty(login)}" th:replace="~{../fragments/login-forms :: forms(web=false, location='/mobile/login')}">
+	<th:block th:if="${#strings.isEmpty(userName)}">
+	<div th:replace="~{../fragments/login-forms :: forms(web=false, location='/mobile/login')}">
 	</div>
+	</th:block>
 
-	<th:block th:unless="${#strings.isEmpty(login)}">
+	<th:block th:unless="${#strings.isEmpty(userName)}">
 	<form method="post" th:action="@{/create-token}">
 		<div class="field is-grouped">
 		  <p class="control">
