+++ phoneblock/src/main/webapp/WEB-INF/templates/en-US/mobile/response.html
@@ -10,8 +10,6 @@
 
 <h1 data-tx="t0001">You are successfully logged in</h1>
 
-<p data-tx="t0002">Your authorization token: <code id="token" th:text="${token}"></code></p>
-
 <p data-tx="t0003">Your login to PhoneBlock was successful. If this page does not close automatically, there may be something wrong with your mobile application.</p>
 
 </div>
