+++ phoneblock/src/main/webapp/WEB-INF/templates/es/mobile/response.html
@@ -10,8 +10,6 @@
 
 <h1 data-tx="t0001">Ha iniciado sesión correctamente</h1>
 
-<p data-tx="t0002">Su token de autorización: <code id="token" th:text="${token}"></code></p>
-
 <p data-tx="t0003">Su acceso a PhoneBlock se ha realizado correctamente. Si esta página no se cierra automáticamente, es posible que haya algún problema con su aplicación móvil.</p>
 
 </div>
