+++ phoneblock/src/main/webapp/WEB-INF/templates/fr/mobile/login.html
@@ -9,15 +9,19 @@
 	<div class="content">
 	<h1 data-tx="t0001">Associer un compte PhoneBlock</h1>
 	
+	<p data-tx="t0005" th:unless="${#strings.isEmpty(userName)}">Tu es connecté à PhoneBlock en tant que <code th:text="${userName}"></code>.</p>
+	
 	<p data-tx="t0002">Ici, tu connectes ton compte PhoneBlock à ton appareil mobile.</p>
 
-	<p data-tx="t0003" th:if="${#strings.isEmpty(login)}">Tu dois d'abord t'inscrire sur PhoneBlock ou créer un compte PhoneBlock. Tu as besoin d'un compte PhoneBlock pour que ton téléphone portable puisse utiliser la base de données PhoneBlock pour filtrer les appels.</p>
+	<p data-tx="t0003" th:if="${#strings.isEmpty(userName)}">Tu dois d'abord t'inscrire sur PhoneBlock ou créer un compte PhoneBlock. Tu as besoin d'un compte PhoneBlock pour que ton téléphone portable puisse utiliser la base de données PhoneBlock pour filtrer les appels.</p>
 	</div>
 
-	<div th:if="${#strings.isEmpty(login)}" th:replace="~{../fragments/login-forms :: forms(web=false, location='/mobile/login')}">
+	<th:block th:if="${#strings.isEmpty(userName)}">
+	<div th:replace="~{../fragments/login-forms :: forms(web=false, location='/mobile/login')}">
 	</div>
+	</th:block>
 
-	<th:block th:unless="${#strings.isEmpty(login)}">
+	<th:block th:unless="${#strings.isEmpty(userName)}">
 	<form method="post" th:action="@{/create-token}">
 		<div class="field is-grouped">
 		  <p class="control">
