+++ phoneblock/src/main/webapp/WEB-INF/templates/fr/mobile/response.html
@@ -10,8 +10,6 @@
 
 <h1 data-tx="t0001">Tu es inscrit(e) avec succès</h1>
 
-<p data-tx="t0002">Ton jeton d'autorisation : <code id="token" th:text="${token}"></code></p>
-
 <p data-tx="t0003">Ton inscription à PhoneBlock a réussi. Si cette page ne se ferme pas automatiquement, il se peut que quelque chose ne fonctionne pas avec ton application mobile.</p>
 
 </div>
