+++ phoneblock/src/main/webapp/WEB-INF/templates/it/mobile/login.html
@@ -9,15 +9,19 @@
 	<div class="content">
 	<h1 data-tx="t0001">Collegare l'account PhoneBlock</h1>
 	
+	<p data-tx="t0005" th:unless="${#strings.isEmpty(userName)}">Siete collegati a PhoneBlock come <code th:text="${userName}"></code>.</p>
+	
 	<p data-tx="t0002">Qui si collega l'account PhoneBlock al dispositivo mobile.</p>
 
-	<p data-tx="t0003" th:if="${#strings.isEmpty(login)}">È necessario prima registrarsi a PhoneBlock o creare un account PhoneBlock. È necessario un account PhoneBlock affinché il telefono cellulare possa utilizzare il database PhoneBlock per filtrare le chiamate.</p>
+	<p data-tx="t0003" th:if="${#strings.isEmpty(userName)}">È necessario prima registrarsi a PhoneBlock o creare un account PhoneBlock. È necessario un account PhoneBlock affinché il telefono cellulare possa utilizzare il database PhoneBlock per filtrare le chiamate.</p>
 	</div>
 
-	<div th:if="${#strings.isEmpty(login)}" th:replace="~{../fragments/login-forms :: forms(web=false, location='/mobile/login')}">
+	<th:block th:if="${#strings.isEmpty(userName)}">
+	<div th:replace="~{../fragments/login-forms :: forms(web=false, location='/mobile/login')}">
 	</div>
+	</th:block>
 
-	<th:block th:unless="${#strings.isEmpty(login)}">
+	<th:block th:unless="${#strings.isEmpty(userName)}">
 	<form method="post" th:action="@{/create-token}">
 		<div class="field is-grouped">
 		  <p class="control">
