+++ phoneblock/src/main/webapp/WEB-INF/templates/it/mobile/response.html
@@ -10,8 +10,6 @@
 
 <h1 data-tx="t0001">L'accesso è stato effettuato con successo</h1>
 
-<p data-tx="t0002">Il suo token di autorizzazione: <code id="token" th:text="${token}"></code></p>
-
 <p data-tx="t0003">Il vostro accesso a PhoneBlock è avvenuto con successo. Se questa pagina non si chiude automaticamente, potrebbe esserci qualcosa di sbagliato nell'applicazione mobile.</p>
 
 </div>
