+++ phoneblock/src/main/webapp/WEB-INF/templates/nb/mobile/response.html
@@ -10,8 +10,6 @@
 
 <h1 data-tx="t0001">Du er vellykket logget inn</h1>
 
-<p data-tx="t0002">Ditt autorisasjonstoken: <code id="token" th:text="${token}"></code>.</p>
-
 <p data-tx="t0003">Innloggingen din til PhoneBlock var vellykket. Hvis denne siden ikke lukkes automatisk, kan det være noe galt med mobilapplikasjonen din.</p>
 
 </div>
