+++ phoneblock/src/main/webapp/WEB-INF/templates/nl/mobile/login.html
@@ -9,15 +9,19 @@
 	<div class="content">
 	<h1 data-tx="t0001">PhoneBlock-account koppelen</h1>
 	
+	<p data-tx="t0005" th:unless="${#strings.isEmpty(userName)}">U bent aangemeld bij PhoneBlock als <code th:text="${userName}"></code>.</p>
+	
 	<p data-tx="t0002">Hier verbindt u uw PhoneBlock-account met uw mobiele apparaat.</p>
 
-	<p data-tx="t0003" th:if="${#strings.isEmpty(login)}">U moet zich eerst registreren bij PhoneBlock of een PhoneBlock-account aanmaken. U hebt een PhoneBlock account nodig zodat uw mobiele telefoon de PhoneBlock database kan gebruiken om oproepen te filteren.</p>
+	<p data-tx="t0003" th:if="${#strings.isEmpty(userName)}">U moet zich eerst registreren bij PhoneBlock of een PhoneBlock-account aanmaken. U hebt een PhoneBlock account nodig zodat uw mobiele telefoon de PhoneBlock database kan gebruiken om oproepen te filteren.</p>
 	</div>
 
-	<div th:if="${#strings.isEmpty(login)}" th:replace="~{../fragments/login-forms :: forms(web=false, location='/mobile/login')}">
+	<th:block th:if="${#strings.isEmpty(userName)}">
+	<div th:replace="~{../fragments/login-forms :: forms(web=false, location='/mobile/login')}">
 	</div>
+	</th:block>
 
-	<th:block th:unless="${#strings.isEmpty(login)}">
+	<th:block th:unless="${#strings.isEmpty(userName)}">
 	<form method="post" th:action="@{/create-token}">
 		<div class="field is-grouped">
 		  <p class="control">
