+++ phoneblock/src/main/webapp/WEB-INF/templates/nl/mobile/response.html
@@ -10,8 +10,6 @@
 
 <h1 data-tx="t0001">U bent succesvol ingelogd</h1>
 
-<p data-tx="t0002">Uw autorisatietoken: <code id="token" th:text="${token}"></code></p>
-
 <p data-tx="t0003">Uw aanmelding bij PhoneBlock is gelukt. Als deze pagina niet automatisch sluit, is er mogelijk iets mis met uw mobiele applicatie.</p>
 
 </div>
