+++ phoneblock/src/main/webapp/WEB-INF/templates/pl/mobile/login.html
@@ -9,15 +9,19 @@
 	<div class="content">
 	<h1 data-tx="t0001">Połącz konto PhoneBlock</h1>
 	
+	<p data-tx="t0005" th:unless="${#strings.isEmpty(userName)}">Jesteś zalogowany do PhoneBlock jako <code th:text="${userName}"></code>.</p>
+	
 	<p data-tx="t0002">Tutaj łączysz swoje konto PhoneBlock z urządzeniem mobilnym.</p>
 
-	<p data-tx="t0003" th:if="${#strings.isEmpty(login)}">Najpierw należy zarejestrować się w usłudze PhoneBlock lub utworzyć konto PhoneBlock. Konto PhoneBlock jest potrzebne, aby telefon komórkowy mógł korzystać z bazy danych PhoneBlock do filtrowania połączeń.</p>
+	<p data-tx="t0003" th:if="${#strings.isEmpty(userName)}">Najpierw należy zarejestrować się w usłudze PhoneBlock lub utworzyć konto PhoneBlock. Konto PhoneBlock jest potrzebne, aby telefon komórkowy mógł korzystać z bazy danych PhoneBlock do filtrowania połączeń.</p>
 	</div>
 
-	<div th:if="${#strings.isEmpty(login)}" th:replace="~{../fragments/login-forms :: forms(web=false, location='/mobile/login')}">
+	<th:block th:if="${#strings.isEmpty(userName)}">
+	<div th:replace="~{../fragments/login-forms :: forms(web=false, location='/mobile/login')}">
 	</div>
+	</th:block>
 
-	<th:block th:unless="${#strings.isEmpty(login)}">
+	<th:block th:unless="${#strings.isEmpty(userName)}">
 	<form method="post" th:action="@{/create-token}">
 		<div class="field is-grouped">
 		  <p class="control">
