+++ phoneblock/src/main/webapp/WEB-INF/templates/pl/mobile/response.html
@@ -10,8 +10,6 @@
 
 <h1 data-tx="t0001">Zostałeś pomyślnie zalogowany</h1>
 
-<p data-tx="t0002">Twój token autoryzacji: <code id="token" th:text="${token}"></code></p>
-
 <p data-tx="t0003">Logowanie do PhoneBlock powiodło się. Jeśli ta strona nie zamknie się automatycznie, być może coś jest nie tak z aplikacją mobilną.</p>
 
 </div>
