+++ phoneblock/src/main/webapp/WEB-INF/templates/sv/mobile/response.html
@@ -10,8 +10,6 @@
 
 <h1 data-tx="t0001">Du är framgångsrikt inloggad</h1>
 
-<p data-tx="t0002">Din behörighetstoken: <code id="token" th:text="${token}"></code>.</p>
-
 <p data-tx="t0003">Din inloggning till PhoneBlock var framgångsrik. Om den här sidan inte stängs automatiskt kan det vara något fel på din mobilapplikation.</p>
 
 </div>
