+++ phoneblock/src/main/webapp/WEB-INF/templates/uk/mobile/login.html
@@ -9,15 +9,19 @@
 	<div class="content">
 	<h1 data-tx="t0001">Прив'язати обліковий запис PhoneBlock</h1>
 	
+	<p data-tx="t0005" th:unless="${#strings.isEmpty(userName)}">Ви увійшли до PhoneBlock як <code th:text="${userName}"></code>.</p>
+	
 	<p data-tx="t0002">Тут ви зв'язуєте свій обліковий запис PhoneBlock з вашим мобільним пристроєм.</p>
 
-	<p data-tx="t0003" th:if="${#strings.isEmpty(login)}">Спочатку ви повинні зареєструватися в PhoneBlock або створити обліковий запис PhoneBlock. Обліковий запис PhoneBlock потрібен для того, щоб ваш мобільний телефон міг використовувати базу даних PhoneBlock для фільтрації дзвінків.</p>
+	<p data-tx="t0003" th:if="${#strings.isEmpty(userName)}">Спочатку ви повинні зареєструватися в PhoneBlock або створити обліковий запис PhoneBlock. Обліковий запис PhoneBlock потрібен для того, щоб ваш мобільний телефон міг використовувати базу даних PhoneBlock для фільтрації дзвінків.</p>
 	</div>
 
-	<div th:if="${#strings.isEmpty(login)}" th:replace="~{../fragments/login-forms :: forms(web=false, location='/mobile/login')}">
+	<th:block th:if="${#strings.isEmpty(userName)}">
+	<div th:replace="~{../fragments/login-forms :: forms(web=false, location='/mobile/login')}">
 	</div>
+	</th:block>
 
-	<th:block th:unless="${#strings.isEmpty(login)}">
+	<th:block th:unless="${#strings.isEmpty(userName)}">
 	<form method="post" th:action="@{/create-token}">
 		<div class="field is-grouped">
 		  <p class="control">
