+++ phoneblock/src/main/webapp/WEB-INF/templates/uk/mobile/response.html
@@ -10,8 +10,6 @@
 
 <h1 data-tx="t0001">Ви успішно увійшли до системи</h1>
 
-<p data-tx="t0002">Ваш токен авторизації: <code id="token" th:text="${token}"></code></p>
-
 <p data-tx="t0003">Ваш вхід до PhoneBlock пройшов успішно. Якщо ця сторінка не закривається автоматично, можливо, з вашим мобільним додатком щось не так.</p>
 
 </div>
