+++ phoneblock/src/main/webapp/WEB-INF/templates/zh-Hans/mobile/login.html
@@ -9,15 +9,19 @@
 	<div class="content">
 	<h1 data-tx="t0001">链接 PhoneBlock 账户</h1>
 	
+	<p data-tx="t0005" th:unless="${#strings.isEmpty(userName)}">您正以 <code th:text="${userName}"></code> 的身份登录 PhoneBlock。</p>
+	
 	<p data-tx="t0002">在这里，您可以将 PhoneBlock 账户与移动设备连接起来。</p>
 
-	<p data-tx="t0003" th:if="${#strings.isEmpty(login)}">您必须首先在 PhoneBlock 注册或创建一个 PhoneBlock 账户。您需要一个 PhoneBlock 帐户，这样您的手机才能使用 PhoneBlock 数据库过滤来电。</p>
+	<p data-tx="t0003" th:if="${#strings.isEmpty(userName)}">您必须首先在 PhoneBlock 注册或创建一个 PhoneBlock 账户。您需要一个 PhoneBlock 帐户，这样您的手机才能使用 PhoneBlock 数据库过滤来电。</p>
 	</div>
 
-	<div th:if="${#strings.isEmpty(login)}" th:replace="~{../fragments/login-forms :: forms(web=false, location='/mobile/login')}">
+	<th:block th:if="${#strings.isEmpty(userName)}">
+	<div th:replace="~{../fragments/login-forms :: forms(web=false, location='/mobile/login')}">
 	</div>
+	</th:block>
 
-	<th:block th:unless="${#strings.isEmpty(login)}">
+	<th:block th:unless="${#strings.isEmpty(userName)}">
 	<form method="post" th:action="@{/create-token}">
 		<div class="field is-grouped">
 		  <p class="control">
