+++ phoneblock/src/main/webapp/WEB-INF/templates/zh-Hans/mobile/response.html
@@ -10,8 +10,6 @@
 
 <h1 data-tx="t0001">您已成功登录</h1>
 
-<p data-tx="t0002">您的授权令牌： <code id="token" th:text="${token}"></code></p>
-
 <p data-tx="t0003">您已成功登录 PhoneBlock。如果此页面没有自动关闭，则可能是您的手机应用程序出了问题。</p>
 
 </div>
